requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
Faker==37.4.2
idna==3.10
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.4.3
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
